// src/components/cron/types.ts
export interface Locale {
  everyText?: string;
  emptyMonths?: string;
  emptyMonthDays?: string;
  emptyMonthDaysShort?: string;
  emptyWeekDays?: string;
  emptyWeekDaysShort?: string;
  emptyHours?: string;
  emptyMinutes?: string;
  emptyMinutesForHourPeriod?: string;
  yearOption?: string;
  monthOption?: string;
  weekOption?: string;
  dayOption?: string;
  hourOption?: string;
  minuteOption?: string;
  rebootOption?: string;
  prefixPeriod?: string;
  prefixMonths?: string;
  prefixMonthDays?: string;
  prefixWeekDays?: string;
  prefixWeekDaysForMonthAndYearPeriod?: string;
  prefixHours?: string;
  prefixMinutes?: string;
  prefixMinutesForHourPeriod?: string;
  suffixMinutesForHourPeriod?: string;
  errorInvalidCron?: string;
  clearButtonText?: string;
  weekDays?: string[];
  months?: string[];
  altWeekDays?: string[];
  altMonths?: string[];
}
// ...
export type PeriodType =
  | "year"
  | "month"
  | "week"
  | "day"
  | "hour"
  | "minute"
  | "reboot"
  | string;
export type AllowEmpty = "always" | "never" | "for-default-value" | string;
export type CronType =
  | "period"
  | "months"
  | "month-days"
  | "week-days"
  | "hours"
  | "minutes"
  | string;
export type LeadingZeroType = "month-days" | "hours" | "minutes" | string;
export type LeadingZero = boolean | LeadingZeroType[];
export type ClockFormat = "24-hour-clock" | "12-hour-clock" | string;
export type ShortcutsType =
  | "@yearly"
  | "@annually"
  | "@monthly"
  | "@weekly"
  | "@daily"
  | "@midnight"
  | "@hourly"
  | "@reboot";
export type Shortcuts = boolean | ShortcutsType[];

// Internal props

export interface FieldProps {
  value?: number[];
  setValue: SetValueNumbersOrUndefined;
  locale: Locale;
  className?: string;
  disabled: boolean;
  readOnly: boolean;
  period: PeriodType;
}
export interface PeriodProps
  extends Omit<FieldProps, "value" | "setValue" | "period"> {
  value: PeriodType;
  setValue: SetValuePeriod;
  shortcuts: Shortcuts;
}
// ...
export type SetValueNumbersOrUndefined = Dispatch<
  SetStateAction<number[] | undefined>
>;
export type SetValuePeriod = Dispatch<SetStateAction<PeriodType>>;